Small change to yesterdays change to ignore keylocking on G1 cameras:
Treat CANON_PS_G1 the same way as CANON_PS_S100 as it apparently also locks
the keys itself when the mode selector wheel is turned to the appropriate
position even if the USB cable is not connected. It is not lacking the
key locking function, it just handles it itself.


git-svn-id: https://svn.code.sf.net/p/gphoto/code/trunk/libgphoto2@3652 67ed7778-7388-44ab-90cf-0a291f65f57c