From: Axel Waggershauser <awagger@web.de>
[..]
5) introduce 3 more macros similar to C_PTP_REP:
  * C_PTP: check ptp-errors, log them with gp_log and translate them
to gp-errors
  * C_PTP_MSG: same as C_PTP but with a custom printf-like error message
  * string
  * C_PTP_REP_MSG: same as C_PTP_REP but with a custom printf-like
error message string

The *_REP_* versions are there to distinguish between errors that
should only be logged and those that should also be reported to the
front-end. It might make sense to reevaluate which errors should be
reported and which should not. I did not change the behaviour here.

The *_MSG versions are there, because I could not find a way to have a
"default" empty message string parameter for the macro. The usage is

    C_PTP (ptp_some_func(x));
    C_PTP_MSG (ptp_some_func(x), "could not set %d", x);

and this would produce a logging line similar to this:

    'ptp_some_func(x)' failed: 'PTP Parameter Not Supported' (0x2006)
    'ptp_some_func(x)' failed: 'could not set 5' (0x2006: 'PTP
Parameter Not Supported')

If one would be happy with writing C_PTP(ptp_some_func(x), ""); all
the time, then the two versions could be merged to one. I preferred
having the extra _MSG version, though.

All C_PTP* macros log a numerical and textual representation of the
PTP_RC_* error code at hand and also log both the calling function and
the called function.

[..]

I dropped the old 'custom' error message in case where it was only
logged (not reported) and only contained a hint to the failing
function, which gets logged automatically with C_PTP now.




git-svn-id: https://svn.code.sf.net/p/gphoto/code/trunk/libgphoto2@15069 67ed7778-7388-44ab-90cf-0a291f65f57c