Add st2205 lookup tables to st2205 camlib
A gphoto user has contacted me about getting what appears to be a new
version of a st2205 picture frame working with gphoto. The st2205 frames
use a rather funky image compression which involves 2 lookup tables:

1) lookup table with 8 byte patterns, a check is done if color
gradients in these blocks are close to a set of fixed patterns, if
they are they the block gets encoded in as a base color + gradient
pattern number

2) These devices do a number of transition effects, these effects
are part of the encoding, the transitions are mainly done to
hide the slowness of the device and the 8x8 blocks are stored
in transition order, so the first block to be shown is the first
one in the data stream, second block to be shown second, etc. This
means we need to know the location where each block will end up
on the screen to compress / decompress. There are a number of
fixed tables for this (1 set of tables per supported resution,
with 1 table per effect number).

So far we've been reading these tables from the firmware image
which can be accessed through the same interface as the picture
memory on these devices. But this new device does not have these
tables in its memory in any recognizable form.

This commits adds the necessary tables directly to the st2205 camlib,
to enable support for this new type of st2205 picture frames.


git-svn-id: https://svn.code.sf.net/p/gphoto/code/trunk/libgphoto2@15102 67ed7778-7388-44ab-90cf-0a291f65f57c