From: Luc Jalbert <axemanlegrand@gmail.com>
2. In camlibs/ptp2/config.c, function
camera_canon_eos_update_capture_target(), if the requested destination is
not the card, there's no need to look through the available destinations to
find out which ID the card is assigned to. It would be an optimization to
just do this if the requested destination is the card.
There's also a theoritical memory access error where the dpd descriptor was
freed and then it was accessed to get the previous value of a field. I
understand that the free operation does not really release any memory
(at least for the value used) but it is probably cleaner to do the read
before the free anyway.



git-svn-id: https://svn.code.sf.net/p/gphoto/code/trunk/libgphoto2@15467 67ed7778-7388-44ab-90cf-0a291f65f57c