Raise errors in ptp2 and add a new PTP_ERROR_NODEVICE error
Currently, in ptp2 (usb), origin of errors are masked because only the
PTP_ERROR_IO error is raised in case of any error.
This fix add translation of gp_errors comming from the port to
meaningful "PTP errors".
An additional "PTP_ERROR_NODEVICE" virtual error is added to allow to
report upstream that a device disappeared and so that the request failed
because of that.