Rework of the ptp event irq manager to raise errors and rearm when needed
This commit rework the ptp2 event (irq) manager for multiple purposes:
- If the callback of an event is called because the async event
encountered an error, report this error upstream.
This allows a call to check_event to detect that a device was removed.
Without it, a call to check event will silentely hide that the
check_event call failed because a device being removed or dead (by
returning a TIMEOUT error) as if everything was ok but there was no
event.

- Re-arm async event requests if less are running than the expected
number of active interrupts transfers

- Use a linked list for the queue of events instead of a memory array
and steal buffers to avoid useless memory copies.
So the code should be faster and more memory efficient.