Fix incomplete EXIF data extraction
If the "GP_FILE_TYPE_EXIF" flag is supplied to get_file_func,
the extracted data is missing two bytes. The value for the maxbytes
variable is read from the APP segment size, which does not account
for the two bytes of the marker that is included in the extracted data.