New set based camlib definition and --with-camlibs handling
This uses proper set operations both for GP_CAMLIB and for handling
the --with-camlibs value. This allows all GP_CAMLIB definitions to
be present at autoreconf time (as m4 code) time while some camlib
definitions being conditional at the time of the configure run (in
the sh code).

The ./configure output is still a bit verbose, but we can reduce that
verbosity later when we have found out that this works well on people's
different systems.

Fixes: https://github.com/gphoto/libgphoto2/issues/467