travis-ci: Move builds to separate script file
The handling of failing build steps has become too involved
to put it into a single command line in a ".travis.yml".

The new build script does a few things differently:
  * No sudo use.
  * Builds for and installs into separate root --prefix
  * Builds as out of tree builds per build

The new things the new build script does:
  * Print config.log and the test-suite.log files if the configure
    or the "make check" stages have failed.
  * Print the version information, both short and verbose.
  * Print a list of the installed files.

The things the new build script does NOT do:
  * Run "make clean" or "make uninstall"

We are keeping the old script lines around for the time
being. I cannot see a difference, but someone more familiar
with "vusb" and "examples/sample-afl" might.