travis-ci: exclude OSX gcc from build matrix
On OSX, 'gcc' is an alias for 'clang', so on OSX, both the
so-called 'gcc' and the 'clang' builds actually use clang.

It would be some work to find out which gcc to maybe install
and then use on OSX, so we disable the 'gcc' builds for now.

If someone wants actual gcc builds on OSX, they can take a
look later at

https://docs.travis-ci.com/user/languages/c/#gcc-on-macos

and experiment with finding out which gcc versions are
actually available by default and available in brew and
use those (after possible first installing them).