Fix when exact requested value is not possible
In the case the requested value is not available on the camera, we got strange behavior before. Shutterspeed got to the end of the range and the other properties jumped 100 steps always around the requested value. 
Now, it is stopped when we go behind the requested value. So as a result the reached value is right the next behind the requested value. 
I think that is the best compromise.

Test: For F-Number: goto 2.0 than try to change to 2.7. Before it will jump 100 times between 2.5 and 2.8. Now it will simply go to 2.8 and stop. 
Shutterspeed similar.