Fix fuji camlib compilation warnings.
fuji.c: Resolve signed-unsigned compare issue. The variable 'i' is used in 'for' loop as array index counter. It is already used as unsigned integer in compares and a negative value would cause a segfault.
library.c: Resolve signed-unsigned compare issue. The variable 'i' is used in 'for' loop in a very simple way and is compared to an unsigned number already.