directory: use statvfs instead of statfs
statfs is a non-portable function available only on few OSes (Linux for
example). Because of this, the storage_info_func provides results only
on Linux. Switch it to statvfs instead: it is a POSIX function, so
portable and supposedly available on any POSIX platform.

Switch the build system bits from statfs to statvs, removing any statfs
leftover. The used APIs of statfs are available as the same when using
statvfs.