camlibs/ptp2: Better cross-platform error handling
WinSock2 doesn't use errno for indicating errors in the network stack.
This breaks the usage of checking errno for network errors as well as
printing them via perror. This introduces the ptpip_perror and
ptpip_get_socket_error function which get the error code and message from
the right pace depending on the platform.