camlibs/ptp2: Add timeout support for connect/read/write socket operations
This switches all sockets to non blocking mode and uses select with a
timeout to limit the maximum wait time for any network operation.

As Windows doesn't use errno for the network ptpip_set_socket_error is
introduced which handles setting the socket error code when a timeout is
encountered so the rest of the camlib can deal handle the error.