Add mode dial positions available on EOS RP or 5Div
I picked C2 and C3 while the existing one is called "Custom"; I think
the naming is a bit ad hoc (full words like "Manual" and "Bulb" vs.
all-caps abbreviations "AV" and "TV" vs. "Fv" which is spelt as-is on
the mode dial).

Tested on EOS RP, 5Div and 5Diii.