Remove rpm packaging
While adding support for "make rpm" appeared to be a good idea in the
early 2000s, it really is not and only clutters up the build system.

At least the libgphoto2_port translation cannot have worked in the
rpm package since at least 5255d802102c10f971dc98f10828451d9922340d
from 2007-10-28, so it is a good guess that this has not really been
in use.