Make CAMLIBS and IOLIBS fallbacks configurable
Make CAMLIBS and IOLIBS fallback locations (where libgphoto2 looks for
camlibs at runtime if the CAMLIBS environment variable is not set, and
where libgphoto2_port looks for iolibs at runtime if the IOLIBS
environment variable is not set) configurable at configure time.

The runtime lookup location DEFAULT_CAMLIBS and DEFAULT_IOLIBS can be
set independently from the installation location camlibdir and iolibdir.

This is to accommodate systems like Windows where it apparently makes
sense to install files to a location at build time which is different
to where the files will eventually end up being used from:

     ./configure \
         --with-camlibdir='/path/to/camlibs' \
         DEFAULT_CAMLIBS='./libgphoto2' \
         --with-iolibdir='/path/to/camlibs' \
         DEFAULT_IOLIBS='./libgphoto2_port'

As this interface is a bit ugly mixing the variables with the
--with-argument, we will improve that in the next commit.