Factor out i18n stuff into separate i18n.h files
Factor out the i18n definitions into i18n.h include files:
One include file each for libgphoto2 and for libgphoto2_port.

This removes the complete

    #ifdef ENABLE_NLS
    #  include <libintl.h>
    #  undef _
    #  define _(String) dgettext (GETTEXT_PACKAGE, String)
    #  ifdef gettext_noop
    #    define N_(String) gettext_noop (String)
    #  else
    #    define N_(String) (String)
    #  endif
    #else
    #  define _(String) (String)
    #  define N_(String) (String)
    #endif

block (which in some occasions defines even more macros)
and replaces it with a single line

    #include "libgphoto2/i18n.h"

for camlibs and libgphoto2 itself and with

    #include "libgphoto2_port/i18n.h"

for iolibs and libgphoto2_port itself.

This gives us two central locations to change the defintions.