Remove linux-hotplug support
The linux-hotplug mechanism has been replaced by udev and hwdb for
quite some time now, so we can remove the support for packaging
libgphoto2 with linux-hotplug.

For a rough idea of the timeline of linux-hotplug being replaced,
the last release of hotplug appears to have been in March of 2004:

https://sourceforge.net/projects/linux-hotplug/files/

and the Fedora 5 (released March 20, 2006) release notes
mention the hotplug package having been removed.