Update libphoto2_port/po/de.po to UTF-8 after all
As the German translation ("de") is configured as "external" at
the Translation Project, there will be no updated `de.po` files
from the Translation Project to overwrite what we have in our
source tree.

So the gphoto project needs to maintain the de.po file independent
from the Translation Project, and updating the file from ISO-8859-1
to UTF-8 is a part of that.