Require at least libtool 2.4.2, use LT_INIT
This gives us a few more features and fixes to rely
on which are present in libtool 2.4.2 (released
2011-10-18), which is older than our autoconf,
automake, and gettext requirements.

Update the libtool initialization to use the more
modern (modern in the sense of libtool 1.9b from
2004-08-29) LT_INIT macro.

We still do not use LTDL_INIT() as that insists on
us shipping libltdl sources in our release tarballs.