Update the checks for jpeg-turbo libjpeg
Update the checks for jpeg-turbo libjpeg to try libjpeg.pc
with PKG_CHECK_MODULES first, and also to allow explicitly
set LIBJPEG_CFLAGS and LIBJPEG_LIBS.

The old mechanism did not allow specifying the location
of libjpeg at all. It only allowed either disabling
libjpeg or using libjpeg from the default include and
linker paths. So this also changes the places in the
Makefiles where we used to add @LIBJPEG@ to the linker to
both link with $(LIBJPEG_LIBS) and compile with
$(LIBJPEG_CFLAGS).

The new configure logic keeps --without-jpeg to force-
disable libjpeg, but updates the detection of libjpeg
to first try the LIBJPEG_* variables if given, otherwise
try PKG_CHECK_MODULES for libjpeg, and only then falls
back to the old way of just trying the default include
and link locations for libjpeg.

Also, this mentions more information in the configure
summary if not building with libjpeg, and it makes sure
to call libjpeg "jpeg-turbo libjpeg" to help people with
finding the proper system package to install.

This is the first time we make it possible to build with
jpeg-turbo libjpeg on e.g. FreeBSD.