Make the compiler warnings work with clang as well
Make the compiler warnings work with clang in addition to gcc.
This came up then the pedantic build tests failed on FreeBSD 13.

To achieve this, we add -Werror=unknown-warning-option to
the CFLAGS for clang to abort immediately on encountering
an unknown warning option. This means we now only add warnings
which the compiler (both clang and gcc) actually supports.

As configure tests that compiling is possible for each
warning flag, this should not introduce compatibility
issues where builds which used to work now do not.

However, you can always opt to not use the generated CFLAGS
by setting CFLAGS to empty when running your build like

    make all check CFLAGS=