Require gettext >= 0.18.3; remove GP_GETTEXT_HACK
This requires gettext version 0.18.3 or later, and it removes
the GP_GETTEXT_HACK which used to generate a po/Makevars file
from the po/Makevers.template.

gettext 0.18.3 is the same gettext version requirement
as libexif uses.

gettext 0.18.3 has been released 2013-07-07, which is
a similar timeframe of the autoconf, automake and libtool
releases we require.

If you ever need to build a post 2021 libgphoto2 from
git on a system with pre 0.18.3 gettext, you can can
always prepare a tarball using "make dist" on a system
with gettext 0.18.3 or later, and build that tarball
on the older machine.

Not generating the po/Makevars file any more means that
we need additional consistency checks for the parts of
po/Makevars which cannot be derived from the Autoconf
standard set of AC_SUBST()ed Makefile.in variables.