Require gettext >= 0.19.1
Require gettext >= 0.19.1 (from 2014-06-10), which both works properly
when building on many cores, and is also the first gettext release
whose po/Makefile.in.in allows disabling rebuilding of the *.po files
on normal (non dist target) builds.

The new po/Makefile.in.in rules fix "make distcheck" consistently
failing to remove ../../../po/libgphoto2-6.pot when running
"make -j$(nproc) distcheck" on a 12 core system:

    rm: cannot remove '../../../po/libgphoto2-6.pot': Permission denied

The "make dist" (and "make distcheck") targets will still update the
po files by default, but that could also be disabled in po/Makevars
if we wanted to separate updating the *.po files for translation from
"make dist" or "make distcheck" at any time in the future.

As gettext-0.19.1 was released almost 8 years ago, this is the newest
build tool release requirement for bulding libgphoto2.

If you ever need to build a 2022 or later libgphoto2 from git on a
system with pre 0.19.1 gettext, you can can always prepare a tarball
running "make dist" on a system with gettext 0.19.1 or later, and
build that tarball on the older machine.