configure: Allow --with-jpeg=yes argument again
Allow the --with-jpeg=yes configure argument again. This
was broken by commit ea00220b6f44a00c2efa792c538f1804ebe183c2.

The GP_LIBJPEG macro in the configure script now handles the following cases:

  --with-jpeg=no
  --without-jpeg
    Do not use libjpeg. Do not autodetect it either.

  --with-jpeg
  --with-jpeg=yes
    Autodetect libjpeg. If not found, abort.

  <no argument given>
  --with-jpeg=auto
  --with-jpeg=autodetect
    Autodetect libjpeg. If found, use it.

However, it does *NOT* handle

  --with-jpeg=/path/to/somewhere

at this time. Set LIBJPEG_CFLAGS and/or LIBJPEG_LIBS
instead to specify which libjpeg to compile with and
link against.

This mostly matches the behaviour from before commit
ea00220b6f44a00c2efa792c538f1804ebe183c2, which was

  --with-jpeg=no
  --without-jpeg
    Do not use libjpeg. Do not autodetect it either.

  --with-jpeg
  --with-jpeg=yes
  --with-jpeg=<any value other than "no">
    Autodetect libjpeg. If found, use it. Otherwise, silently
    ignore the user's wish to build with libjpeg.

  <no argument given>
    Autodetect libjpeg. If found, use it.

If someone is running configure with --with-jpeg=yes and
libjpeg cannot be used, we can assume that building without
using libjpeg is not the outcome the user expects.

Fixes: https://github.com/gphoto/libgphoto2/issues/759