Allow clang compile with initializer overrides
This catches clang in a clang specific C preprocessor conditional
before clang gets a chance to mess up at pretending to be gcc.

The camlibs/pentax/js0n.c state machine tables use a lot of both
GNU style range designators [ a ... b ], and later overrides some
of the values in the range. This is not very pure C, but relatively
effient to code, so we can accept silencing some warnings here.

Fixes: https://github.com/gphoto/libgphoto2/issues/761