gh workflow: Use parallel make for builds
Use parallel make for builds. We use the GNU coreutils'
"nproc" utility to determine the number of cores available
for us to use. We then tell make to launch up to that number
of processes (-j$(nproc)) in parallel, but only if the system
load is below $(nproc) (-l$(nproc)).

That should make reasonably good use of however many cores github
makes available to us, while not overloading the build server
or the build server VM.