gh workflow: work around libgd-dev dependency problem
Work around libgd-dev dependency issue apparently caused by a 3rd party
repository with PHP and nginx and related stuff like libgd.

So we remove nginx and libgd3 to get rid of the relevant 3rd party
packages which pull in their libgd, and only then install libgd-dev
which then comes from Ubuntu proper and works.

As we need neither nginx nor php, this should work for libgphoto2.

This appears to be a reoccuring problem on github runners, so keeping
this workaround in place looks like the way to go for the future.

See also https://github.com/actions/runner-images/issues/6153
and recursively follow the linked issues and comments for more
information.