Fix -Wcalloc-transposed-args warnings
This switches the arguments of calloc() in a few cases such that
the sizeof() does not happen in the nmemb first argument.

An example warning would be
warning: 'calloc' sizes specified with 'sizeof' in the earlier argument and not in the later argument [-Wcalloc-transposed-args]

This is the variant of commit f97c27efe9f1d50d31879657569405808099b54e
with the typo removed which broke compilation.