improve API of ptp_unpack_string and ptp_unpack_uint[16|32]_array
Those functions get passed an offset value into the data buffer. This
offset value needed to be updated on the calling side after the call with
internal knowledge of what exactly the unpack function itself was doing.
Now the offset becomes an in/out parameter and in updated internally. This
substantially reduces the complexity of the calling code.

Furthermore, the return value is now an error indicator and the array
length gets set via a pointer.

This further fixes a few places where there was a bug in the bounds
checking or no bounds checking at all.