make camera_trigger_canon_eos_capture "synchronous"
This is a draft of a change along the lines of the discussion in
https://github.com/gphoto/libgphoto2/issues/968#issuecomment-2337742476
meant to be able to discuss something like that further.

After realizing that there is no event sent from the camera to reliably
detect the moment the camera is ready again, I went back to the approach
that I successfully used for the last 10 years in production.

This allows me to execute the following command on my R8 to capture a
total of 4 images in under 2s.

gphoto2 --set-config capturetarget=1
  --set-config aeb="+/- 3"
  --set-config-index drivemode=1
  --set-config shutterspeed=1/500
  --set-config ownername=400
  --trigger-capture
  --set-config aeb=off
  --set-config-index drivemode=0
  --set-config ownername=1
  --trigger-capture

The use of "ownername" is obviously not meant to stay, it was just a quick
hack to enable me to pass an sleep_ms parameter into
camera_trigger_capture from the command line.