ptp2: remove unused PTPCanon_Property members size and data
PTPDeviceProperty data of EOS cameras is not polled but rather pro-actively
sent by the camera via a PropValueChanged event. The unpack function used
to interpret the data and unpack it into the PTPDevicePropDesc struct
but also copied the raw event data block into an extra member (data, size)
of the special special PTPCanon_Property struct. This data is then never
looked at again.

This patch removes those data members and is a first step to completely
eliminate the special PTPCanon_Property struct.