ptp2: drop the "Supported" string in some PTPDeviceInfo members
The PTPDeviceInfo struct contains information about the Operations, Events,
DeviceProperties, ImageFormats and CaptureFormats. With 3 of those 5 the
variables contained the additional "Supported" string, which does not add
any discriminating value and only makes for very long and therefore less
readable code lines and it is inconsistent.

Also use the term "DeviceProps" here as well for brevity and consistency.