ptp2: replace nrofsomearray pattern with somearray_len pattern
The "_len"-suffix to denote the length of an array of values is more
readable than the "nrof"-prefix and also used 3 times more often in the
rest of the code base.

Also replace the params->deviceproperties name with params->dpd_cache to
make it clearer what it actually is: an array of cached DevicePropDesc
structs.