ptp2: move 'oid' variable into PTPObjectInfo as (Object) 'Handle'
The (Object) Handle member is not part of the "ObjectInfo dataset" in the
specification but added here to avoid having to carry it around by
separate means throughout the code base, as it is regularly required in
combination. While the spec always refers to this item as "ObjectHandle",
the term "Handle" is only used in this context, so we omit the
"Object"-prefix for brevity.

Switching from oid/ObjectID, which is nowere to be found the spec,
increases consistency with the spec and the rest of the code base.

This allows to remove the struct _PTPCanon_New_Object from the codebase.

@msmeissn I added 2 TODO comments for you take a close look, pointing out:
 * a memory leak that I did not want to touch because I can't test it
 * a presumed bug fix where newobject was 0 when passed to ptp_getobject