ptp2: replace PTPObjectHandles and PTPStorageIDs with generic array
This adds a set of macros implementing a generic array or list of TYPE.
This is basically a TYPE* pointer and a length integer. This structure
together with the typical use-cases repeats regularly throughout the
codebase. It raises the level of abstraction and improves code
readabilty. I'm a c++ developer and miss my STL ;).

I'm starting here with PTPObjectHandles and PTPStorageIDs as those were
already modeling an array abstraction. There are other arrays of stuff
that can benefit from this later.