ptp2: camera_wait_for_event: Olympus: Only allocate path buffer when needed
Previously the memory for the object path was allocated whether there was a valid object handle returned by the ObjectAdded event or not.
In case there was no non-zero object handle the allocated memory was neither passed through to the event handler to free it nor was it freed internally.