Sony prioritymode fix
Sony mode 3 cameras have this setting, but it is uint8 instead of int8. Tested on A7R4/5 and A1M2.

This setting allows overriding camera dials via PTP. Most Sony cameras will have an exposure program (e.g M, P, S, A, etc) that can be set via a dial on the camera, normally this cannot be set by PTP, prioritymode allows the exposure program to be changed. Depending on the camera focus mode and drive mode may also be unlocked with this property.