Add support for A1M2
To allow shutter speed to be set on A1M2, the old alternative shutter speed property had to be removed as this property is reused for something else on A1M2 (and other newer cameras). This is ok as the normal shutter speed property is available on cameras that need to use the alternative shutter speed property, and we can redirect to use the alternative shutter speed property as needed (Tested on A7R4 using mode 2).