ptp2: fix memory leak in camera_init
ptp_list_folder can be called with a NULL as the last parameter. Otherwise
the returned array must be freed by the caller. But here we don't use the
list of handles anyway.

Also instead of excluding the code for Sony, only perform it for Canon,
which was the reason this was re-introduced.