ptp2: re-introduce root caching in ptp_list_folder
This patch does 3 things:
1) It brings back the caching of root entries in ptp_list_folder that was
   mentioned in #1177.
2) It silences an uninformative and annoying stream of "error 0x2002"
   messages on the console with the 5Dm2 that was mentioned e.g. in #1028
3) It simplifies the ptp_list_folder_eos code by removing the unused
   internal handling of querying multiple storages in a row.

camera_init now always calls ptp_list_folder for each storage once. This
should also take care of the Canon EOS 1500D that seems to hang if not
done. The code has been tested on a 5Ds and a R5m2 and works with 2 cards
inserted at the same time.